Fix showtime quá khứ + Trang Chính sách linh động =))

+++ app/Http/Controllers/Admin/ShowtimeController.php
@@ -40,6 +40,8 @@ class ShowtimeController extends Controller
     public function index(Request $request)
     {
         // Giá trị mặc định
+        // $timeNow = now()->format('Y-m-d H:i:s');
+        // dd($timeNow);
         $user = Auth::user();
         if ($user->cinema_id == "") {
             $defaultBranchId = Branch::where('is_active', 1)->first()?->id ?? null;
@@ -301,6 +303,7 @@ class ShowtimeController extends Controller
 
     public function edit(Showtime $showtime)
     {
+        $timeNow = now();
         $seatShowtimes = SeatShowtime::where('showtime_id', $showtime->id)->pluck('status', 'id')->all();
         // dd($seatShowtime);
         $statusSeat = true;
@@ -311,24 +314,30 @@ class ShowtimeController extends Controller
             }
         }
         if ($statusSeat) {
-            $showtimes = Showtime::with(['room', 'movieVersion'])->get();
+            if (!$timeNow->greaterThan($showtime->start_time)) {
+                $showtimes = Showtime::with(['room', 'movieVersion'])->get();
 
-            $movies = Movie::where('is_active', '1')->get();
-            $user = auth()->user();
-            if ($user->cinema_id == "") {
-                $rooms = Room::where('is_active', '1')->with(['cinema'])->first('id')->get();
+                $movies = Movie::where('is_active', '1')->get();
+                $user = auth()->user();
+                if ($user->cinema_id == "") {
+                    $rooms = Room::where('is_active', '1')->with(['cinema'])->first('id')->get();
+                } else {
+                    $rooms = Room::with('typeRoom', 'seats')->where('is_active', '1')->where('cinema_id', $user->cinema_id)->get();
+                }
+                $movieVersions = MovieVersion::all();
+                $cinemas = Cinema::where('is_active', '1')->with(['branch'])->first('id')->get();
+                $branches = Branch::where('is_active', '1')->get();
+
+                $movieDuration = $showtime->movie->duration;
+
+
+                $cleaningTime = Showtime::CLEANINGTIME;
+                return view(self::PATH_VIEW . __FUNCTION__, compact('movies', 'rooms', 'movieVersions', 'cinemas', 'cleaningTime', 'branches', 'showtime', 'movieDuration'));
             } else {
-                $rooms = Room::with('typeRoom', 'seats')->where('is_active', '1')->where('cinema_id', $user->cinema_id)->get();
+                return redirect()
+                    ->route('admin.showtimes.index')
+                    ->with('error', 'Bạn không được sửa suất chiếu trong quá khứ!');
             }
-            $movieVersions = MovieVersion::all();
-            $cinemas = Cinema::where('is_active', '1')->with(['branch'])->first('id')->get();
-            $branches = Branch::where('is_active', '1')->get();
-
-            $movieDuration = $showtime->movie->duration;
-
-
-            $cleaningTime = Showtime::CLEANINGTIME;
-            return view(self::PATH_VIEW . __FUNCTION__, compact('movies', 'rooms', 'movieVersions', 'cinemas', 'cleaningTime', 'branches', 'showtime', 'movieDuration'));
         } else {
             return redirect()
                 ->route('admin.showtimes.index')
@@ -352,7 +361,7 @@ class ShowtimeController extends Controller
             $endTime = $startTime->copy()->addMinutes($movieDuration + $cleaningTime);
 
             $dataShowtimes = [
-                'cinema_id' => isset($request->cinema_id) ? $request->cinema_id : $user->cinema_id,
+                // 'cinema_id' => isset($request->cinema_id) ? $request->cinema_id : $user->cinema_id,
                 'room_id' => $request->room_id,
                 'format' => $typeRoom->name . ' ' . $movieVersion->name,
                 'movie_version_id' => $request->movie_version_id,
@@ -381,6 +390,8 @@ class ShowtimeController extends Controller
     {
 
         try {
+
+            $timeNow = now();
             $seatShowtimes = SeatShowtime::where('showtime_id', $showtime->id)->pluck('status', 'id')->all();
             // dd($seatShowtime);
             $statusSeat = true;
@@ -391,13 +402,19 @@ class ShowtimeController extends Controller
                 }
             }
             if ($statusSeat) {
-                $showtime->delete();
+                if (!$timeNow->greaterThan($showtime->start_time)) {
+                    $showtime->delete();
 
-                return redirect()
-                    ->route('admin.showtimes.index')
-                    ->with('success', 'Xóa thành công!');
+                    return redirect()
+                        ->route('admin.showtimes.index')
+                        ->with('success', 'Xóa thành công!');
+                } else {
+                    return redirect()
+                        ->route('admin.showtimes.index')
+                        ->with('error', 'Bạn không được sửa suất chiếu trong quá khứ!');
+                }
             } else {
-                return  redirect()
+                return redirect()
                     ->route('admin.showtimes.index')
                     ->with('error', 'Xóa không thành công! Suất chiếu này đã/đang có người đặt!');
             }
